pyproject.toml
[project]
name = "talculate"
version = "0.1.0"
description = "A programmer oriented tui calculator. simple keys. minimal ui"
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
authors = [
  { name = "nooneknows", email = "157999171+nooneknowspeter@users.noreply.github.com" },
]
dependencies = [
  "pyperclip (>=1.9.0,<2.0.0)",
  "pyyaml (==6.0.2)",
  "rich (>=14.1.0,<15.0.0)",
  "textual (==3.7.0)",
  "xdg (==6.0.0)",
]

[project.scripts]
talc = "talculate.main:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true
packages = [{ include = "talculate", from = "src" }]

